+++ tomcat/src/main/java/org/apache/controller/RegisterController.java
@@ -6,8 +6,8 @@ import java.util.Optional;
 import java.util.UUID;
 import org.apache.exception.InvalidRequestException;
 import org.apache.http.Cookie;
-import org.apache.http.HttpRequest;
 import org.apache.http.HttpResponse;
+import org.apache.http.request.HttpRequest;
 import org.apache.http.value.HttpMethod;
 import org.apache.session.Session;
 import org.apache.session.SessionManager;
