+++ tomcat/src/main/java/org/apache/controller/StaticFileController.java
@@ -11,8 +11,8 @@ import java.nio.file.Paths;
 import java.util.List;
 import org.apache.exception.InvalidRequestException;
 import org.apache.exception.RequestProcessingException;
-import org.apache.http.HttpRequest;
 import org.apache.http.HttpResponse;
+import org.apache.http.request.HttpRequest;
 import org.apache.http.value.ContentType;
 import org.apache.http.value.HttpHeader;
 import org.apache.http.value.StatusCode;
